Add missing logging tests
